# perf(map_based_prediction): create a fence LineString layer and use rtree query