--- a/perception/autoware_map_based_prediction/include/map_based_prediction/map_based_prediction_node.hpp
+++ b/perception/autoware_map_based_prediction/include/map_based_prediction/map_based_prediction_node.hpp
@@ -39,6 +39,7 @@
 #include <visualization_msgs/msg/marker_array.hpp>
 
 #include <lanelet2_core/Forward.h>
+#include <lanelet2_core/LaneletMap.h>
 #include <lanelet2_routing/Forward.h>
 #include <lanelet2_traffic_rules/TrafficRules.h>
 
@@ -177,6 +178,9 @@ private:
 
   // Crosswalk Entry Points
   lanelet::ConstLanelets crosswalks_;
+
+  // Fences
+  lanelet::LaneletMapUPtr fence_layer_{nullptr};
 
   // Parameters
   bool enable_delay_compensation_;